Ignore init arg length check for Hotbar
If they use a hardcoded message as their Hotbar message in config.js then we want to allow them to use that by simply enabling the Hotbar with the command !hotbar. This will default to the message set forth in config.js and save the user the hassle of retyping that message every time they want to enable it and use that specific message.

Note: This backup of the original message set forth in config.js is held in memory only. To reset this backup you need to reboot the server.