Various changes and fix for OP Password
Most changes are minor and mean nothing to the community. However a bug fix is included in this patch that now allows the user to change their password in their configuration file and to properly utilize that password to update themselves with a new hash and salt so they can gain proper access. Before this patch, the command was bugged and would not allow you to change your password and gain access. Ergo, forcing you to go back to your old passord. This is now resolved.