Fix TypeError in tick event for worldborder
'typeId' returns null while loading into the world if worldborder is enabled. Its null because there are no chunks to read when checking the blocks around the player. Using the nullish coalescing operator we can return its right-hand side operand when its left-hand side operand is null or undefined. In this case, we go ahead and return the blocks as air so it continues the code in the script as to be expected until the player and chunks are finally loaded.