Update package.json to add @minecraft/server-ui override for @minecraft/server
The `server-ui` depends on the latest stable version of  `server` even if beta, but they are not cross compatible so we tell it which version to target by overriding the default to enforce compatibility.

If you are still facing errors after pulling in this patch then delete the node_modules folder in the root directory and the `src` directory. Pull dependencies again with the command `npm i`.